Extract cross-tenant claim validation into shared JsClaimsUtil to remove duplication
