+++ components/authentication-framework/org.wso2.carbon.identity.application.authentication.framework/src/main/java/org/wso2/carbon/identity/application/authentication/framework/config/model/graph/js/openjdk/nashorn/JsOpenJdkNashornClaims.java
@@ -21,12 +21,12 @@ package org.wso2.carbon.identity.application.authentication.framework.config.mod
 import org.apache.commons.lang.StringUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-import org.wso2.carbon.context.PrivilegedCarbonContext;
 import org.wso2.carbon.identity.application.authentication.framework.ApplicationAuthenticator;
 import org.wso2.carbon.identity.application.authentication.framework.config.ConfigurationFacade;
 import org.wso2.carbon.identity.application.authentication.framework.config.model.ExternalIdPConfig;
 import org.wso2.carbon.identity.application.authentication.framework.config.model.StepConfig;
 import org.wso2.carbon.identity.application.authentication.framework.config.model.graph.js.AbstractJSContextMemberObject;
+import org.wso2.carbon.identity.application.authentication.framework.config.model.graph.js.JsClaimsUtil;
 import org.wso2.carbon.identity.application.authentication.framework.context.AuthenticationContext;
 import org.wso2.carbon.identity.application.authentication.framework.exception.UserIdNotFoundException;
 import org.wso2.carbon.identity.application.authentication.framework.internal.FrameworkServiceDataHolder;
@@ -40,7 +40,6 @@ import org.wso2.carbon.identity.claim.metadata.mgt.ClaimMetadataHandler;
 import org.wso2.carbon.identity.claim.metadata.mgt.exception.ClaimMetadataException;
 import org.wso2.carbon.identity.core.IdentityClaimManager;
 import org.wso2.carbon.identity.core.util.IdentityTenantUtil;
-import org.wso2.carbon.identity.core.util.IdentityUtil;
 import org.wso2.carbon.idp.mgt.IdentityProviderManagementException;
 import org.wso2.carbon.idp.mgt.IdentityProviderManager;
 import org.wso2.carbon.user.api.UserRealm;
@@ -62,7 +61,6 @@ import java.util.Optional;
 public class JsOpenJdkNashornClaims extends AbstractJSContextMemberObject implements AbstractOpenJdkNashornJsObject {
 
     private static final Log LOG = LogFactory.getLog(JsOpenJdkNashornClaims.class);
-    private static final String SAAS_ENABLE_CROSS_TENANT_OPERATIONS = "SaaS.EnableCrossTenantOperations";
     private String idp;
     private boolean isRemoteClaimRequest;
     private int step;
@@ -452,45 +450,6 @@ public class JsOpenJdkNashornClaims extends AbstractJSContextMemberObject implem
 
     protected boolean isAuthenticatedUserInCurrentTenant() {
 
-        if (authenticatedUser == null) {
-            return false;
-        }
-
-        if (isSaasApp(getContext()) && isSaaSCrossTenantOperationsEnabled()) {
-            return true;
-        }
-
-        if (IdentityTenantUtil.isTenantQualifiedUrlsEnabled()) {
-            return StringUtils.equals(authenticatedUser.getTenantDomain(),
-                    PrivilegedCarbonContext.getThreadLocalCarbonContext().getTenantDomain());
-        }
-
-        AuthenticationContext context = getContext();
-        if (context == null || StringUtils.isBlank(context.getTenantDomain())) {
-            LOG.warn("Unable to determine the tenant domain from the authentication context. " +
-                    "Hence user tenant domain validation is considered as failed.");
-            return false;
-        }
-        return StringUtils.equals(authenticatedUser.getTenantDomain(), context.getTenantDomain());
-    }
-
-    private static boolean isSaaSCrossTenantOperationsEnabled() {
-
-        String value = IdentityUtil.getProperty(SAAS_ENABLE_CROSS_TENANT_OPERATIONS);
-        if (StringUtils.isBlank(value)) {
-            return false;
-        }
-        return Boolean.parseBoolean(value);
-    }
-
-    private static boolean isSaasApp(AuthenticationContext context) {
-
-        if (context == null || context.getSequenceConfig() == null
-                || context.getSequenceConfig().getApplicationConfig() == null
-                || context.getSequenceConfig().getApplicationConfig().getServiceProvider() == null) {
-            LOG.debug("Unable to determine if the application is a SaaS app. Treating as non-SaaS app.");
-            return false;
-        }
-        return context.getSequenceConfig().getApplicationConfig().getServiceProvider().isSaasApp();
+        return JsClaimsUtil.isAuthenticatedUserInCurrentTenant(authenticatedUser, getContext());
     }
 }
